normie/requirements.txt
# Requirements for Normie AI Projects

# Local Chatbot
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
chromadb>=0.3.0

# Chess Pattern Analyzer
python-chess>=1.999
pandas>=2.0.0
numpy>=1.24.0

